refactored listener to its own class
